ceph: delete terminating osd pods even if not out
When an OSD pod is stuck in terminating state, there is no need
to check if the OSD is marked as out before force deleting.
As long as the osd is down we will force delete the pod.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 3c6bae56b70df9ef440936dde253f379f9ae2ea5)