ceph: osd on pvc node assignment is node selector
An OSD on a PVC when portable=false is assigned to a node
with a node selector. The same node assignment is expected
for the lifetime of the cluster. On subsequent reconciles,
the operator was looking up the node assignment from the
nodeName on the pod spec, which is not set if the pod is down.
The operator needs to retrieve the assignment from the
deployment spec node selector.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 65377cbe633c0ea2ff670bcd977ab04d79c1618e)
(cherry picked from commit 3ce9db491a2961c23dc46eb14adcdee7723c2ea2)