ceph: require rgw pods to be running to create object users
The object user controller was intending to wait for rgw pods to be running
before creating a user, however, after querying for the pods it didn't check
if any pods were actually running. Creating the user before the object
store was running causes the default rgw pools to potentially be created
by ceph instead of allowing Rook to create the pools with the expected
settings in the CR. Now we wait properly for the rgw to start before
creating the object users.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>