ceph: ignore unscheduled ceph daemons while cleaning up
Before cleaning up the cluster, we wait for all the daemon pods to be cleaned up. This fails when
a daemon is in pending state and has no NodeName. This PR ignores daemon pods that are not scheduled on any node.

Signed-off-by: Santosh Pillai <sapillai@redhat.com>
(cherry picked from commit 753bdb3526ab56e6c6ce51215a6b97099bda7df9)