ceph: cancel on-going orchestration on any CR update
If the CephCluster CR is updated let's cancel any on-going
orchestration. This will handle the situation where the spec is
misconfigured and OSD are crashlooping.
Prior to this patch, we would have to wait for the provisioning to
timeout, which will leave the cluster in a bad state for 10min or more.

Now the CR changes are picked up right away and the OSDs will deploy
normally.

Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit 40607c11169a5b8da5ee3c5ba48ade0253e57324)