ceph: unset the prometheus server_addr before updating mgr
The server_addr was previously set to the pod IP, but now the prometheus binding
is expected to use the default binding to localhost. In new installs, the
binding will be to localhost, but upgrades were retaining the setting to
binding to the pod IP. Since the pod IP can change when the pod is restarted,
the binding will fail. Thus, we unset the server_addr binding so the default
localhost will be used.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>