ceph: add timeout in mgr init container
The mgr init containers run a ceph command that may get
stuck if the mon quorum is down. The liveness probe is not
in effect during the init containers, thus the mgr pod
may stay hung indefinitely when this happens. To ensure the
mgr does not hang, we add a bash timeout to the ceph command
so it will restart and try again in case mon quorum is restored.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>